- Move Attribute class to a separate file and update references
