+++ src/app/submission/submission-shared/model/index.ts
@@ -3,4 +3,5 @@ export * from './submission-to-pagetab.utils';
 export * from './pagetab-to-submission.utils';
 export * from './pagetab';
 export * from './submission';
+export * from './submission/submission.model.attribute';
 export * from './templates';
