+++ src/app/submission/submission-shared/model/submission/submission.model.ts
@@ -1,5 +1,6 @@
 import { zip } from 'fp-ts/lib/Array';
 import { nextId } from './submission.model.counter';
+import { Attribute } from './submission.model.attribute';
 import { NameAndValue, Tag } from '../model.common';
 import {
     ColumnType,
@@ -8,41 +9,11 @@ import {
     FieldType,
     SectionType,
     SubmissionType,
-    ValueType,
-    ValueTypeFactory
+    ValueType
 } from '../templates';
 
 interface SubmissionSection {
     subsections: Sections
-}
-
-export class Attribute {
-    readonly id: string;
-
-    constructor(
-        private _name: string = '',
-        readonly valueType: ValueType = ValueTypeFactory.DEFAULT,
-        readonly displayType: DisplayType = DisplayType.Optional,
-        readonly isTemplateBased: boolean = false,
-        readonly dependencyColumn: string = '',
-        readonly uniqueValues: boolean = false
-    ) {
-        this.id = `attr_${nextId()}`;
-    }
-
-    get name(): string {
-        return this._name;
-    }
-
-    set name(name: string) {
-        if (this.canEditName && this._name !== name) {
-            this._name = name;
-        }
-    }
-
-    get canEditName(): boolean {
-        return !this.isTemplateBased;
-    }
 }
 
 export class AttributeValue {
